+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/elevator/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package holds the {@link io.github.thebusybiscuit.slimefun4.implementation.items.elevator.ElevatorPlate} and any
+ * related classes to making the elevator work.
+ */
+package io.github.thebusybiscuit.slimefun4.implementation.items.elevator;