+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 **Table of contents**
 
-- [Release Candidate 19 (TBD)](#release-candidate-19-tbd)
+- [Release Candidate 19 (11 Jan 2021)](#release-candidate-19-11-jan-2021)
 - [Release Candidate 18 (03 Dec 2020)](#release-candidate-18-03-dec-2020)
 - [Release Candidate 17 (17 Oct 2020)](#release-candidate-17-17-oct-2020)
 - [Release Candidate 16 (07 Sep 2020)](#release-candidate-16-07-sep-2020)
@@ -24,20 +24,39 @@
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
 
-## Release Candidate 19 (TBD)
+## Release Candidate 20 (TBD)
+
+#### Additions
+
+#### Changes
+
+#### Fixes
+* Fixed elevator floor order
+
+## Release Candidate 19 (11 Jan 2021)
 
 #### Additions
 * Added Bee Armor (1.15+ only)
+* (API) Added AndroidFarmEvent
 
 #### Changes
 * Performance optimizations for Cargo networks
 * Removed an old version of bStats
+* CraftBukkit is officially no longer supported, Slimefun will now be disabled on old builds of CraftBukkit
+* Removed the deprecated ItemManipulationAPI for BlockMenus
+* Removed the "Written Book" variant of the Slimefun Guide
+* The Elevator has an Inventory menu now
 
 #### Fixes
 * Fixed a couple of compatibility issues with ItemsAdder
 * Fixed #2575
 * Fixed ghost blocks to some extent (ghost blocks will now drop and be replaced)
 * Fixed #2636 (hotfix)
+* Fixed #2647
+* Fixed #2664
+* Fixed #2655
+* Fixed /sf timings --verbose not working correctly
+* Fixed #2675
 
 ## Release Candidate 18 (03 Dec 2020)